pequenas melhorias gerais
