+++ package.json
@@ -49,6 +49,7 @@
     "react": "18.2.0",
     "react-admin": "^4.10.3",
     "react-dom": "18.2.0",
+    "react-icons": "^4.8.0",
     "reflect-metadata": "^0.1.13",
     "typeorm": "^0.3.7",
     "typescript": "5.0.2"
